Improved build script

+++ htdocs/build.sh
@@ -1,6 +1,26 @@
 #!/bin/bash
-cd ..
-svn checkout svn://svn.berlios.de/websane/htdocs
+
+if [ -d "htdocs" ]
+then
+	svn update htdocs
+else 
+	svn checkout svn://svn.berlios.de/websane/htdocs
+	chmod a+x htdocs/build.sh
+fi
+
 xsltproc --xinclude htdocs/xslt/builder.xsl htdocs/xml/site.xml
-ln -s htdocs/style `pwd`/style
-ln -s htdocs/images `pwd`/images
+
+if [ ! -L "style" ] 
+then
+	ln -s htdocs/style `pwd`/style
+fi
+if [ ! -L "images" ] 
+then
+	ln -s htdocs/images `pwd`/images
+fi
+if [ ! -L "build.sh" ]
+then
+	rm build.sh
+	ln htdocs/build.sh ./build.sh
+fi
+
